[core] rearrange service and pubsub headers into folders.
