+++ app/rec/rec_client_core/src/frame.h
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2019 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@
 #include <string>
 #include <chrono>
 #include <ecal/ecal_time.h>
-#include <ecal/ecal_callback.h>
+#include <ecal/pubsub/types.h>
 
 namespace eCAL
 {
