+++ app/sys/sys_cli/src/commands/command.h
@@ -1,6 +1,6 @@
 /* ========================= eCAL LICENSE =================================
  *
- * Copyright (C) 2016 - 2020 Continental Corporation
+ * Copyright (C) 2016 - 2025 Continental Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,7 +26,7 @@
 #include <memory>
 #include <ecalsys/ecal_sys.h>
 
-#include <ecal/ecal_client.h>
+#include <ecal/service/client.h>
 
 #ifdef _MSC_VER
 #pragma warning(push)
