+++ ecal/core/include/ecal/ecal_registration.h
@@ -25,7 +25,8 @@
 #pragma once
 
 #include <ecal/ecal_os.h>
-#include <ecal/ecal_types.h>
+#include <ecal/pubsub/types.h>
+#include <ecal/service/types.h>
 
 #include <cstdint>
 #include <functional>
