+++ ecal/core/include/ecal/service/server.h
@@ -18,7 +18,7 @@
 */
 
 /**
- * @file   ecal_server.h
+ * @file   server.h
  * @brief  eCAL server interface
 **/
 
@@ -28,9 +28,7 @@
 #include <ecal/ecal_namespace.h>
 #include <ecal/ecal_os.h>
 
-#include <ecal/ecal_callback.h>
-#include <ecal/ecal_service_info.h>
-#include <ecal/ecal_types.h>
+#include <ecal/service/types.h>
 
 #include <memory>
 #include <string>
