+++ ecal/core/include/ecal/v5/ecal_publisher.h
@@ -29,7 +29,7 @@
 #include <ecal/v5/ecal_callback.h>
 #include <ecal/ecal_deprecate.h>
 #include <ecal/ecal_os.h>
-#include <ecal/ecal_payload_writer.h>
+#include <ecal/pubsub/payload_writer.h>
 #include <ecal/ecal_config.h>
 #include <ecal/ecal_types.h>
 #include <ecal/config/publisher.h>
