+++ ecal/core/src/service/ecal_clientgate.h
@@ -23,6 +23,7 @@
 
 #pragma once
 
+#include <ecal/ecal_namespace.h>
 #include "ecal_def.h"
 #include "serialization/ecal_struct_sample_registration.h"
