+++ ecal/core/src/v5/service/ecal_service_client.cpp
@@ -27,7 +27,7 @@
 
 #include <ecal/ecal.h>
 #include <ecal/v5/ecal_client.h>
-#include <ecal/ecal_service_info.h>
+#include <ecal/service/types.h>
 
 namespace eCAL
 {
