+++ ecal/core/src/v5/service/ecal_service_client_impl.h
@@ -26,7 +26,7 @@
 #include <ecal/ecal_deprecate.h>
 #include <ecal/ecal_namespace.h>
 #include <ecal/ecal_os.h>
-#include <ecal/ecal_client.h>
+#include <ecal/service/client.h>
 #include <ecal/v5/ecal_callback.h>
 
 #include <map>
