+++ ecal/samples/cpp/misc/timer/src/timer.cpp
@@ -18,7 +18,7 @@
 */
 
 #include <ecal/ecal.h>
-#include <ecal/ecal_publisher.h>
+#include <ecal/pubsub/publisher.h>
 
 #include <iostream>
 #include <functional>
