+++ ecal/tests/cpp/pubsub_test/src/pubsub_test.cpp
@@ -20,8 +20,8 @@
 #include <chrono>
 #include <cstddef>
 #include <ecal/ecal.h>
-#include <ecal/ecal_publisher.h>
-#include <ecal/ecal_subscriber.h>
+#include <ecal/pubsub/publisher.h>
+#include <ecal/pubsub/subscriber.h>
 
 #include <atomic>
 #include <functional>
