+++ ecal/tests/cpp/pubsub_test/src/pubsub_test_shm.cpp
@@ -19,8 +19,8 @@
 
 #include <cstddef>
 #include <ecal/ecal.h>
-#include <ecal/ecal_publisher.h>
-#include <ecal/ecal_subscriber.h>
+#include <ecal/pubsub/publisher.h>
+#include <ecal/pubsub/subscriber.h>
 
 #include <atomic>
 #include <functional>
